change table usage
